+++ app/javascript/channels/messagerie_channel.js
@@ -1,0 +1,16 @@
+import consumer from "./consumer";
+
+const initChatroomCable = () => {
+  const messagesContainer = document.getElementById('messages');
+  if (messagesContainer) {
+    const id = messagesContainer.dataset.chatroomId;
+
+    consumer.subscriptions.create({ channel: "ChatroomChannel", id: id }, {
+      received(data) {
+        messagesContainer.insertAdjacentHTML('beforeend', data);
+      }
+    });
+  }
+}
+
+export { initChatroomCable };
