+++ app/javascript/packs/application.js
@@ -31,4 +31,5 @@ document.addEventListener("turbolinks:load", () => {
   initAutocomplete();
   initMapbox();
   initChatroomCable();
+  initMessagerieCable();
 });
