+++ app/views/messageries/show.html.erb
@@ -13,4 +13,5 @@
 
 <%= simple_form_for [ @messagerie, @message ], remote: true do |f| %>
   <%= f.input :content, label: false, placeholder: "Message ##{@messagerie.user.code}" %>
+  <%= f.submit "Post" %>
 <% end %>
