+++ app/views/messages/_message.html.erb
@@ -1,0 +1,7 @@
+<div class="message-container" id="message-<%= message.id %>">
+  <i class="author">
+    <span><%= message.user.code %></span>
+    <small><%= message.created_at.strftime("%a %b %e at %l:%M%p") %></small>
+  </i>
+  <p><%= message.content %></p>
+</div>
