+++ test/controllers/messages_controller_test.rb
@@ -1,0 +1,7 @@
+require "test_helper"
+
+class MessagesControllerTest < ActionDispatch::IntegrationTest
+  # test "the truth" do
+  #   assert true
+  # end
+end
